pyproject.toml
[project]
name = "hotboard"
version = "0.1.0"
description = "World hot topics aggregator - Auto scrape trending content from multiple platforms"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "HotBoard", url = "https://github.com/yourusername/hotboard"}]
keywords = ["scraper", "aggregator", "news", "trending", "hot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.27.0",
    "feedparser>=6.0.0",
    "newspaper3k>=0.2.8",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
    "aiofiles>=23.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
